reworked to be simpler to use and async
all fonts are just rendered to 512 size and shrunk to the size you want.  This makes choosing an output size quick and saves having to messa bout with a sdf range. It always just renders to 512 and does a 32 pixel SDF range,  then bicubic scales down to whatever output size you want up to 64 pixels high.