go back to a low memory solution
Tried converting a 22,000 character sumsun font and it was running out of memory with all the large sdf caching.
going to just go back to a preview only mode so it only allocates temporary memory as it creates each SDF.  this makes each SDF creation very slow, but at least it doesn't crash on huge files now.